[compiler]: Move module namespace up one level out of sema to decouple the systems.
[support]: Decouple diagnostics from the compiler library.
