+++ lib/compiler/src/ast/expressions/call.cpp
@@ -2,13 +2,16 @@
 
 #include "ast/expressions/call.hpp"
 
-#include "ast/expressions/enum.hpp" // IWYU pragma: keep
+// IWYU pragma: begin_keeps
+#include "ast/expressions/enum.hpp"
 #include "ast/expressions/function.hpp"
-#include "ast/expressions/identifier.hpp"    // IWYU pragma: keep
-#include "ast/expressions/scope_resolve.hpp" // IWYU pragma: keep
-#include "ast/expressions/struct.hpp"        // IWYU pragma: keep
+#include "ast/expressions/identifier.hpp"
+#include "ast/expressions/scope_resolve.hpp"
+#include "ast/expressions/struct.hpp"
 #include "ast/expressions/type.hpp"
-#include "ast/expressions/union.hpp" // IWYU pragma: keep
+#include "ast/expressions/union.hpp"
+// IWYU pragma: end_keeps
+
 #include "ast/visitor.hpp"
 
 namespace porpoise::ast {
