+++ lib/compiler/src/module/memory_loader.cpp
@@ -1,6 +1,6 @@
-#include "sema/module/memory_loader.hpp"
+#include "module/memory_loader.hpp"
 
-namespace porpoise::sema::mod {
+namespace porpoise::mod {
 
 auto MemoryLoader::add(const std::filesystem::path& path, const std::string& content) -> void {
     const auto normalized = normalize(path);
@@ -13,11 +13,11 @@ auto MemoryLoader::load(const std::filesystem::path& path) -> Result<std::string
     assert(normalized);
     auto it = files_.find(normalized->string());
     if (it == files_.end()) {
-        return make_sema_err(
+        return make_mod_err(
             fmt::format("Could not find path '{}' in virtual file system", path.string()),
             Error::PATH_DOES_NOT_EXIST);
     }
     return it->second;
 }
 
-} // namespace porpoise::sema::mod
+} // namespace porpoise::mod
