+++ lib/compiler/tests/module/mock.inc
