+++ lib/compiler/tests/module/test_memory_loader.cpp
@@ -1,12 +1,12 @@
 #include <catch2/catch_test_macros.hpp>
 
-#include "sema/module/memory_loader.hpp"
+#include "module/memory_loader.hpp"
 
 namespace porpoise::tests {
 
 TEST_CASE("Correct add/load cycle") {
-    const std::string       expected_content = "This is a file";
-    sema::mod::MemoryLoader loader;
+    const std::string expected_content = "This is a file";
+    mod::MemoryLoader loader;
     loader.add("mock", expected_content);
 
     const auto content = loader.load("mock");
@@ -15,7 +15,7 @@ TEST_CASE("Correct add/load cycle") {
 }
 
 TEST_CASE("Overwriting entries in the VFS") {
-    sema::mod::MemoryLoader loader;
+    mod::MemoryLoader loader;
     loader.add("mock", "This is the bad content");
 
     const std::string expected_content = "This is the good content";
@@ -27,12 +27,12 @@ TEST_CASE("Overwriting entries in the VFS") {
 }
 
 TEST_CASE("Missing path in VFS") {
-    sema::mod::MemoryLoader loader;
-    const auto              result = loader.load("mock");
+    mod::MemoryLoader loader;
+    const auto        result = loader.load("mock");
     REQUIRE_FALSE(result);
 
-    const sema::Diagnostic expected{"Could not find path 'mock' in virtual file system",
-                                    sema::Error::PATH_DOES_NOT_EXIST};
+    const mod::Diagnostic expected{"Could not find path 'mock' in virtual file system",
+                                   mod::Error::PATH_DOES_NOT_EXIST};
     CHECK(result.error() == expected);
 }
